refactor: extract recurring contributions into focused module (#679)
